+++ recipes/login_defs.rb
@@ -21,6 +21,8 @@ template "/etc/login.defs" do
   variables(
     :pass_max_days => node["stig"]["login_defs"]["pass_max_days"],
     :pass_min_days => node["stig"]["login_defs"]["pass_min_days"],
-    :pass_warn_age => node["stig"]["login_defs"]["pass_warn_age"]
+    :pass_warn_age => node["stig"]["login_defs"]["pass_warn_age"],
+    :pass_min_len => node["stig"]["login_defs"]["pass_min_len"],
+    :fail_delay => node["stig"]["login_defs"]["fail_delay"]
   )
 end